renamed autoemplete componet file
